+++ src/consumer/consumer.js
@@ -8,7 +8,7 @@ import chalk from 'chalk';
 import format from 'string-format';
 import partition from 'lodash.partition';
 import symlinkOrCopy from 'symlink-or-copy';
-import { locateConsumer, pathHasConsumer } from './consumer-locator';
+import { locateConsumer, pathHasConsumer, pathHasBitMap } from './consumer-locator';
 import { ConsumerAlreadyExists, ConsumerNotFound, NothingToImport, MissingDependencies } from './exceptions';
 import { Driver } from '../driver';
 import DriverNotFound from '../driver/exceptions/driver-not-found';
@@ -308,7 +308,7 @@ export default class Consumer {
         dist,
         saveDependenciesAsComponents
       });
-      if (installNpmPackages) await this.installNpmPackages(componentsAndDependenciesBitJson);
+      if (installNpmPackages) await this.installNpmPackages(componentsAndDependenciesBitJson, verbose);
     }
     if (componentsFromBitMap.length) {
       componentsAndDependenciesBitMap = await this.scope.getManyWithAllVersions(componentsFromBitMap, cache);
@@ -360,6 +360,7 @@ export default class Consumer {
     withPackageJson?: boolean = true,
     writeBitDependencies?: boolean = false,
     force?: boolean = false,
+    verbose?: boolean = false,
     dist?: boolean = false,
     conf?: boolean = false,
     installNpmPackages?: boolean = true,
@@ -381,7 +382,7 @@ export default class Consumer {
       dist,
       saveDependenciesAsComponents
     });
-    if (installNpmPackages) await this.installNpmPackages(componentsWithDependencies);
+    if (installNpmPackages) await this.installNpmPackages(componentsWithDependencies, verbose);
     return { dependencies: componentsWithDependencies };
   }
 
@@ -420,6 +421,7 @@ export default class Consumer {
       withPackageJson,
       writeBitDependencies,
       force,
+      verbose,
       dist,
       conf,
       installNpmPackages,
@@ -513,7 +515,7 @@ export default class Consumer {
         withPackageJson,
         origin,
         consumer: this,
-        writeBitDependencies,
+        writeBitDependencies: writeBitDependencies || !componentWithDeps.component.dependenciesSavedAsComponents, // when dependencies are written as npm packages, they must be written in package.json
         dependencies: componentWithDeps.dependencies,
         componentMap
       });
@@ -983,6 +985,9 @@ export default class Consumer {
   static async load(currentPath: string): Promise<Consumer> {
     const projectPath = locateConsumer(currentPath);
     if (!projectPath) return Promise.reject(new ConsumerNotFound());
+    if (!pathHasConsumer(projectPath) && pathHasBitMap(projectPath)) {
+      await Consumer.create(currentPath).then(consumer => consumer.write());
+    }
     const scopeP = Scope.load(path.join(projectPath, BIT_HIDDEN_DIR));
     const bitJsonP = ConsumerBitJson.load(projectPath);
     return Promise.all([scopeP, bitJsonP]).then(
@@ -1174,7 +1179,12 @@ export default class Consumer {
     });
   }
 
-  async installNpmPackages(componentsWithDependencies: ComponentWithDependencies[], verbose: boolean): Promise<*> {
+  async installNpmPackages(
+    componentsWithDependencies: ComponentWithDependencies[],
+    verbose: boolean = false
+  ): Promise<*> {
+    // if dependencies are installed as bit-components, go to each one of the dependencies and install npm packages
+    // otherwise, if the dependencies are installed as npm packages, npm already takes care of that
     const componentsWithDependenciesFlatten = R.flatten(
       componentsWithDependencies.map((oneComponentWithDependencies) => {
         return oneComponentWithDependencies.component.dependenciesSavedAsComponents
@@ -1185,12 +1195,11 @@ export default class Consumer {
     loader.start(BEFORE_INSTALL_NPM_DEPENDENCIES);
     const results = await Promise.all(
       componentsWithDependenciesFlatten.map((component) => {
-        const packagesToInstall =
-          component._bitDependenciesPackages && !component.dependenciesSavedAsComponents
-            ? Object.assign(component._bitDependenciesPackages, component.packageDependencies)
-            : component.packageDependencies;
-        if (R.isEmpty(packagesToInstall)) return Promise.resolve();
-        return npmClient.install(packagesToInstall, { cwd: component.writtenPath });
+        // don't pass the packages to npmClient.install function.
+        // otherwise, it'll try to npm install the packages in one line 'npm install packageA packageB' and when
+        // there are mix of public and private packages it fails with 404 error.
+        // passing an empty array, results in installing packages from the package.json file
+        return npmClient.install([], { cwd: component.writtenPath }, verbose);
       })
     );
     loader.stop();